Module Endo.Make

The implementation of smart constructors.

Parameters

module P : Param

Signature

val shifted : P.level -> P.Shift.t -> P.level

shifted s l is the smarter version of Syntax.Endo.shifted that collapses multiple displacements, representing the level l shifted by the displacement s.

raises Invalid_argument
When it (directly or indirectly) attempts to shift the top level.

val top : P.level

top is Syntax.Endo.top, the additional top level for convenience.