Module type Endo.S

The signature of smart constructors.

type shift

The displacement algebra.

type level

The type that embeds levels.

val shifted : level -> shift -> level

shifted s l is the smarter version of Syntax.Endo.shifted that collapses multiple displacements, representing the level l shifted by the displacement s.

raises Invalid_argument
When it (directly or indirectly) attempts to shift the top level.

val top : level

top is Syntax.Endo.top, the additional top level for convenience.