Module type Free.S

The signature of smart constructors.

type shift

The displacement algebra.

type var

The type of level variables.

type level = (shift, var) Syntax.free

The type of freely generated levels.

val var : var -> level

var is Syntax.Free.var, representing the variable level v.

val shifted : level -> shift -> level

shifted s l is the smarter version of Syntax.Endo.shifted that collapses multiple displacements, representing the level l shifted by the displacement s.

raises Invalid_argument
When it (directly or indirectly) attempts to shift the top level.

val top : level

top is Syntax.Endo.top, the additional top level for convenience.