Module Endo.Make

The implementation of smart constructors.

Parameters

module P : Param

Signature

type level = P.level

The type that embeds levels.

val const : int -> level

const is Syntax.Endo.const

val shifted : level -> P.Shift.t -> level

Smarter version of Syntax.Endo.shifted that collapses multiple shifting.

val top : level

top is Syntax.Endo.top

val simplify : level -> level

simplify l collapses multiple shifting operators and removes useless ones. (For example, shifted top is still top.)

val dissect : level -> level * P.Shift.t option

dissect l is a helper function that separate the shifting operators from a level (if any). If the level is shifted multiple times, the shifting operators are composed into one.