Module type Free.S

The signature of smart constructors.

module Shift : Shift.S

The class of shifting operators.

type var

The type of level variables.

type level = ( Shift.t, var ) Syntax.free

The type of freely generated levels.

val var : var -> level

var is Syntax.Free.var

include Endo.S with module Shift := Shift and type level := level

val shifted : level -> Shift.t -> level

Smarter version of Syntax.Endo.shifted that collapses multiple shifting.

val top : level

top is Syntax.Endo.top

val simplify : level -> level

simplify l collapses multiple shifting operators and removes useless ones. (For example, shifted top is still top.)

val dissect : level -> level * Shift.t option

dissect l is a helper function that separate the shifting operators from a level (if any). If the level is shifted multiple times, the shifting operators are composed into one.