Module Mugenjou.Shift

module type S = sig ... end

Common interface of classes of shifting operators.

module Crude : S

Conor McBride's crude stratification that contains only fun i -> i + n.

type crude = Crude.t

module Linear : sig ... end

Generalized Crude that allows scaling.

type linear = Linear.t

module LinearPostInc : sig ... end

Another generalized Crude that allows scaling and post increments.

type lpi = LinearPostInc.t