Module MultiShift.LiftBoundedSemilattice

An instance of Expr inspired by Agda and Lean.

Parameters

module Var : StructuredType.OrderedType

module Base : BoundedSemilattice.S

Signature

include Expr with type var := Var.t

include StructuredType.PartiallyOrderedType

include StructuredType.EqualityType

type t

The type.

val equal : t -> t -> bool

equal x y checks whether x and y are equivalent.

val dump : Stdlib.Format.formatter -> t -> unit

Ugly printer.

val lt : t -> t -> bool

lt x y checks if x is strictly less than y. Note that trichotomy fails for general partial orders.

val leq : t -> t -> bool

leq x y checks if x is less than or equal to y. Note that trichotomy fails for general partial orders.

val var : Var.t -> t

var v is the variable v as an expression.

val subst : ( Var.t -> t ) -> t -> t

subst f e substitutes every variable v in e with f v.

val bot : t

bot is the minimum expression.

val join : t -> t -> t

join x y is the maximum value of x and y.

val const : Base.t -> t

const n represents the constant level n.

val act : Base.t -> t -> t

act s l applies s to the expression l.