Module type MultiShift.Expr

Level expressions with variables for Multi

include StructuredType.PartiallyOrderedType

include StructuredType.EqualityType

type t

The type.

val equal : t -> t -> bool

equal x y checks whether x and y are equivalent.

val dump : Stdlib.Format.formatter -> t -> unit

Ugly printer.

val lt : t -> t -> bool

lt x y checks if x is strictly less than y. Note that trichotomy fails for general partial orders.

val leq : t -> t -> bool

leq x y checks if x is less than or equal to y. Note that trichotomy fails for general partial orders.

type var

Type of variables.

val var : var -> t

var v is the variable v as an expression.

val subst : ( var -> t ) -> t -> t

subst f e substitutes every variable v in e with f v.