Module Shift.Constant

Augmentation with constants.

Parameters

module Base : S

module Const : RightAction with type act := Base.t

Signature

include S

include StructuredType.PartiallyOrderedType

include StructuredType.EqualityType

type t

The type.

val equal : t -> t -> bool

equal x y checks whether x and y are equivalent.

val dump : Stdlib.Format.formatter -> t -> unit

Ugly printer.

val lt : t -> t -> bool

lt x y checks if x is strictly less than y. Note that trichotomy fails for general partial orders.

val leq : t -> t -> bool

leq x y checks if x is less than or equal to y. Note that trichotomy fails for general partial orders.

val id : t

id is the identity (no shifting).

val is_id : t -> bool

is_id s checks whether s is the identity. It is equivalent to equal id s, but potentially faster.

val compose : t -> t -> t

compose s1 s2 composes the operators s1 and s2. Note that Foo^s1^s2 in McBride's notation is understood as compose (compose ... s2) s1 with the reversed order.

val act : Base.t -> t

act s represents actions.

val const : Const.t -> t

const s represents constants.

val to_either : t -> ( Base.t, Const.t ) Stdlib.Either.t

to_either convert an element to a value of type Either.t