Module Shift.Fractal

Parameters

module Base : S

Signature

include S

type t

The type of shifting operators.

val id : t

id is the identity (no shifting).

val equal : t -> t -> bool

equal x y checks whether x and y are equivalent.

val is_id : t -> bool

is_id s checks whether s is the identity. It is equivalent to equal id s, but potentially faster.

val lt : t -> t -> bool

lt x y checks if x is strictly less than y. Note that trichotomy fails for general partial orders.

val leq : t -> t -> bool

leq x y checks if x is less than or equal to y. Note that trichotomy fails for general partial orders.

val compose : t -> t -> t

compose s0 s1 composes the operators s0 and s1. Note that Foo^k1^k2 in McBride's notation is understood as compose (compose ... k2) k1.

val dump : Stdlib.Format.formatter -> t -> unit

Ugly printer.

val embed : Base.t -> t

val push : Base.t -> t -> t