Module Shift.Gapped

Generalized Crude that allows gaps.

The motivation is to enable insertion of universe levels between any two consecutive levels, which is something the function in Crude cannot do. The idea is to introduce "gaps" in these functions, or in other words, to generalize crude shifting operators in Crude to piecewise crude shifting operators. Technically, we consider all strictly monotone functions f such that f(i) = i + n when i >= k for some finite k, where "gaps" may occur for levels i < k, which means in-between universe levels can be introduced as levels within the gaps.

For example, the following shifting operator has three gaps: 0-9, 11-199, and 201-999:

f(0) = 10
f(1) = 200
f(2) = 1000
f(3) = 1001
f(4) = 1002
...
f(i) = i + 997 (for i >= 3)
...

New universe levels that will be made available by this function are 0, 1, ..., 8, 9, 11, 12, ..., 198, 199, 201, 202, ..., 998, and 999.

The class of gapped functions is closed under identity and composition, and is strictly larger than the class in Crude: the functions in Crude are the functions with k = 0, or equivalently the functions with at most one gap in the very beginning.

include S

type t

The type of shifting operators.

val id : t

id is the identity (no shifting).

val const : int -> t

const n is the constant shifting fun i -> i + n.

raises Invalid_argument
if n < 0.

val is_id : t -> bool

is_id s checkes whether s is the identity.

val equal : t -> t -> bool

equal x y checks whether x and y are the same operator.

val lt : t -> t -> bool

lt x y checks if x is strictly less than y. Note that trichotomy in general fails for shifting operators.

val le : t -> t -> bool

le x y checks if x is less than or equal to y. Note that trichotomy in general fails for shifting operators.

val compose : t -> t -> t

compose s0 s1 composes the operators s0 and s1, in the texual order.

val dump : Stdlib.Format.formatter -> t -> unit

Ugly printer.

val of_steps : int list -> t

of_steps can be used to create any shifting operator in this class. of_steps [s0; s1; s2; ...; sn] is the following function:

f(0) = s0
f(1) = s0 + s1
f(2) = s0 + s1 + s2
...
f(i) = s0 + s1 + s2 + ... + si
...
f(n) = s0 + s1 + s2 + ... + sn
f(n+1) = s0 + s1 + s2 + ... + sn + 1
...
f(n+k) = s0 + s1 + s2 + ... + sn + k
...

As a special case, of_steps [] is equivalent to id.

raises Invalid_argument
if s0 < 0 or si < 1 for any i > 0.

val of_skipped : int list -> t

of_skipped can be used to create any shifting operator in this class. of_skipped l gives the shifting operator that skips the levels in l. The numbers in l should be non-negative and strictly increasing. For example, of_skipped [0; 1] is equivalent const 2 and of_steps ~init:2 [] because the first two levels are skipped. of_skipped [3] is equivalent to of_steps ~init:0 [1;1;2] because both correspond to the function

f(0) = 0
f(1) = 1
f(2) = 2
f(3) = 4 (skipping 3)
f(4) = 5
...

of_skipped [0; 1; 2; 4; 5; 7] is equivalent to of_steps ~init:3 [3; 2] because both correspond to the function

f(0) = 3 (skipping 0, 1, and 2)
f(1) = 6 (skipping 4 and 5)
f(2) = 8 (skipping 7)
f(3) = 9
...

raises Invalid_argument
if any element in l is not non-negative or if l is not strictly increasing.

val of_prefix : int list -> t

of_prefix can be used to create any shifting operator in this class. of_prefix l gives the shifting operator whose prefix is l and whose gaps are clearly marked in l. As a special case, of_prefix [] is equivalent to id.

For example, of_prefix [2] is equivalent to of_skipped [0; 1], const 2, and of_steps ~init:2 []. of_prefix [0; 1; 2; 4] is equivalent to of_skipped [3] and of_steps ~init:0 [1;1;2]. of_prefix [3; 6; 8] is equivalent to of_skipped [0; 1; 2; 4; 5; 7] and of_steps ~init:3 [3; 2].

raises Invalid_argument
if any element in l is not non-negative or if l is not strictly increasing.