Module Shift.LinearPostInc

Translation, scaling, and post increments.

The motivation is to enable insertion of universe levels between any two consecutive levels, which is something the function in Crude cannot do. The idea is to introduce scaling and post increments so that we are considering the higher-order functions fun f i -> f (i * n1 + n0) + n2 instead of just fun i -> i + n.

The function g from levels to levels is embedded into the class of higher-order functions as fun f i -> f (g i).

include OrderedMonoid

type t

The type of shifting operators.

val id : t

id is the identity (no shifting).

val equal : t -> t -> bool

equal x y checks whether x and y are equivalent.

val is_id : t -> bool

is_id s checks whether s is the identity. It is equivalent to equal id s, but potentially faster.

val lt : t -> t -> bool

lt x y checks if x is strictly less than y. Note that trichotomy in general fails for partial orders.

val le : t -> t -> bool

le x y checks if x is less than or equal to y. Note that trichotomy in general fails for partial orders.

val compose : t -> t -> t

compose s0 s1 composes the operators s0 and s1, in the textual order.

val dump : Stdlib.Format.formatter -> t -> unit

Ugly printer.

val trans : int -> t

trans n is the translation fun f i -> f (i + n) (the embedding of fun i -> i + n).

raises Invalid_argument
if n < 0.

val scale : int -> t

scale n is the scaling fun f i -> f (i * n) (the embedding of fun i -> i * n).

val postinc : int -> t

postinc n is the post increment fun f i -> f i + n.