Module type Shift.MultiExpr

Level expressions with variables for MultiExpr

type var

Type of variables.

type t

Type of expressions.

val var : var -> t

var v is the variable v as an expression.

val subst : ( var -> t ) -> t -> t

subst f e substitutes every variable v in e with f v.

val equal : t -> t -> bool

equal x y returns true if x and y are equivalent experssions.

val lt : t -> t -> bool

lt x y returns true if x is strictly less than y.

val leq : t -> t -> bool

leq x y returns true if x is less than or equal to y.

val dump : Stdlib.Format.formatter -> t -> unit

Ugly printer for expressions.