Module type Shift.S

Common interface of classes of shifting operators.

type t

The type of shifting operators.

val id : t

id is the identity (no shifting).

val translate : int -> t

translate n is the translation fun i -> i + n.

raises Invalid_argument
if n < 0.

val is_degen : t -> [ `Id | `Const of int ] option

is_id s checkes whether s is the identity or represents a constant function.

val equal : t -> t -> bool

equal x y checks whether x and y are the same operator.

val lt : t -> t -> bool

lt x y checks if x is strictly less than y. Note that trichotomy in general fails for shifting operators.

val le : t -> t -> bool

le x y checks if x is less than or equal to y. Note that trichotomy in general fails for shifting operators.

val compose : t -> t -> t

compose s0 s1 composes the operators s0 and s1, in the texual order.

val apply : int -> t -> int

apply i s applies the shifting operator s to the constant level i.

val dump : Stdlib.Format.formatter -> t -> unit

Ugly printer.