Module Mugen.ShiftWithJoin

Displacement algebras with joins

module type Semilattice = sig ... end

A displacement ordered monoid with joins.

module type BoundedSemilattice = sig ... end

A displacement ordered monoid with joins and a bottom element.

module Int : sig ... end

The displacement algebra of integers.

module Nat : sig ... end

The displacement algebra of natural numbers.

module BinaryProduct (X : Semilattice) (Y : Semilattice) : sig ... end

The binary product of two displacement semilattices.

module LexicalBinaryProduct (X : BoundedSemilattice) (Y : BoundedSemilattice) : sig ... end

The binary product of two bounded displacement semilattices, but with the lexicographical order.

module InfiniteProduct (Base : Semilattice) : sig ... end

The infinite product of a displacement semilattice.