Parameter Lexicographic.Y

include Semilattice

include Shift.S

type t

The type.

val equal : t -> t -> bool

equal x y checks whether x and y are equivalent.

val dump : Stdlib.Format.formatter -> t -> unit

Ugly printer.

val lt : t -> t -> bool

lt x y checks if x is strictly less than y. Note that trichotomy fails for general partial orders.

val leq : t -> t -> bool

leq x y checks if x is less than or equal to y. Note that trichotomy fails for general partial orders.

val id : t

id is the unit.

val is_id : t -> bool

is_id s checks whether s is the unit. It is equivalent to equal id s, but potentially faster.

val compose : t -> t -> t

compose s1 s2 composes the operators s1 and s2. Note that Foo^s1^s2 in McBride's notation is understood as compose (compose ... s2) s1 with the reversed order.

val join : t -> t -> t

join x y is the maximum of x and y.

val bot : t

bot is the minimum value.