Module type StructuredType.PartiallyOrderedTypeWithRightAction

include PartiallyOrderedType

include EqualityType

type t

The type.

val equal : t -> t -> bool

equal x y checks whether x and y are equivalent.

val dump : Stdlib.Format.formatter -> t -> unit

Ugly printer.

val lt : t -> t -> bool

lt x y checks if x is strictly less than y. Note that trichotomy fails for general partial orders.

val leq : t -> t -> bool

leq x y checks if x is less than or equal to y. Note that trichotomy fails for general partial orders.

type act

the underlying type of an algebra

val act : t -> act -> t

the right action