Module Syntax.Endo

Stupid constructors for endo.

type ('s, 'a) t = ('s, 'a) endo =

| Shifted of 'a * 's

| Top

A family of polynomial endofunctors ('s, -) t indexed by the type of displacements 's.

val shifted : 'a -> 's -> ('s, 'a) t

shifted l s is Shifted (l, s), representing the level l shifted by the displacement s.

val top : ('s, 'a) t

top is Top, the additional top level for convenience.

val dump : (Stdlib.Format.formatter -> 's -> unit) -> (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> ('s, 'a) t -> unit

dump dump_s dump_a is the ugly printer for levels, where dump_s is the printer for displacements and dump_a is the printer for inner sub-expressions.