Module Syntax.Free

Stupid constructors for free.

type ('s, 'v) t = ('s, 'v) free =

| Level of ('s, ('s, 'v) free) endo

| Var of 'v

The free monad ('s, -) t on the endofunctor ('s, -) endo indexed by the type of displacements 's.

val shifted : ('s, 'v) t -> 's -> ('s, 'v) t

shifted l s is Level (Shifted (l, s)), representing the level l shifted by the displacement s.

val top : ('s, 'v) t

top is Top, the additional top level for convenience.

val var : 'v -> ('s, 'v) t

var v is Var v, representing the variable level v.

val dump : (Stdlib.Format.formatter -> 's -> unit) -> (Stdlib.Format.formatter -> 'v -> unit) -> Stdlib.Format.formatter -> ('s, 'v) t -> unit

dump dump_s dump_v is the ugly printer for levels, where dump_s is the printer for displacements and dump_v is the printer for variables.