mugen: Universe Levels

Links

🔰 Quickstart tutorial

📔 API reference

What is "mugen"?

"mugen" is the transliteration of "無限" in Japanese, possibly a learned borrowing of "無限" from Chinese. It literally means "without limits", and is widely used in anime for the obvious reason. It is fitting in the context of universe polymorphism.