mugen: Universe Levels

Links

ğŸ”° Quickstart tutorial

ğŸ“” API reference

What is "mugen"?

"mugen" is the transliteration of "ç„¡é™" in Japanese, possibly a learned borrowing of "ç„¡é™" from Chinese. It literally means "without limits", and is widely used in anime for the obvious reason. It is fitting in the context of universe polymorphism.